--- modules/apps/portal-workflow/portal-workflow-kaleo-definition-impl/src/main/java/com/liferay/portal/workflow/kaleo/definition/internal/export/NodeExporterRegistry.java
@@ -1,58 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.workflow.kaleo.definition.internal.export;
-
-import com.liferay.osgi.service.tracker.collections.map.ServiceReferenceMapperFactory;
-import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
-import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
-import com.liferay.portal.workflow.kaleo.definition.NodeType;
-import com.liferay.portal.workflow.kaleo.definition.exception.KaleoDefinitionValidationException;
-import com.liferay.portal.workflow.kaleo.definition.export.NodeExporter;
-
-import org.osgi.framework.BundleContext;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
-
-/**
- * @author Michael C. Han
- */
-@Component(service = NodeExporterRegistry.class)
-public class NodeExporterRegistry {
-
-	public NodeExporter getNodeExporter(NodeType nodeType)
-		throws KaleoDefinitionValidationException {
-
-		return _serviceTrackerMap.getService(nodeType);
-	}
-
-	@Activate
-	protected void activate(BundleContext bundleContext) {
-		_serviceTrackerMap = ServiceTrackerMapFactory.openSingleValueMap(
-			bundleContext, NodeExporter.class, null,
-			ServiceReferenceMapperFactory.create(
-				bundleContext,
-				(nodeExporter, emitter) -> emitter.emit(
-					nodeExporter.getNodeType())));
-	}
-
-	@Deactivate
-	protected void deactivate() {
-		_serviceTrackerMap.close();
-	}
-
-	private ServiceTrackerMap<NodeType, NodeExporter> _serviceTrackerMap;
-
-}
